Remove ThemeProvider and move themes to chart context
