+++ projects/js-packages/charts/src/providers/chart-context/index.ts
@@ -4,3 +4,4 @@ export { useChartId } from './hooks/use-chart-id';
 export { useChartRegistration } from './hooks/use-chart-registration';
 export { useGlobalChartsTheme } from './hooks/use-global-charts-theme';
 export type { GlobalChartsContextValue, ChartRegistration } from './types';
+export * from './themes';
