+++ projects/js-packages/charts/src/providers/index.ts
@@ -1,2 +1,1 @@
 export * from './chart-context';
-export * from './theme';
