--- projects/js-packages/charts/src/providers/theme/index.ts
@@ -1,2 +1,0 @@
-export { ThemeProvider, useChartTheme } from './theme-provider';
-export { defaultTheme, jetpackTheme, wooTheme } from './themes';
