--- projects/js-packages/charts/src/providers/theme/stories/index.stories.tsx
@@ -1,168 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import { ThemeProvider, jetpackTheme, wooTheme } from '../.';
-import { LineChart, BarChart, PieSemiCircleChart } from '../../../.';
-import { medalCountsData as barSampleData } from '../../../stories/sample-data';
-
-const meta: Meta< typeof ThemeProvider > = {
-	title: 'JS Packages/Charts/Themes',
-	component: ThemeProvider,
-	parameters: {
-		layout: 'centered',
-	},
-};
-
-export default meta;
-type Story = StoryObj< typeof ThemeProvider >;
-
-const sampleData = [ barSampleData[ 0 ], barSampleData[ 1 ], barSampleData[ 2 ] ];
-
-const lineSampleData = [
-	{
-		group: 'Line 1',
-		label: 'Line 1',
-		data: [
-			{ date: new Date( '2024-01-01' ), value: 10, label: 'Jan 1' },
-			{ date: new Date( '2024-01-02' ), value: 20, label: 'Jan 2' },
-			{ date: new Date( '2024-01-03' ), value: 15, label: 'Jan 3' },
-			{ date: new Date( '2024-01-04' ), value: 25, label: 'Jan 4' },
-			{ date: new Date( '2024-01-05' ), value: 30, label: 'Jan 5' },
-		],
-	},
-	{
-		group: 'Line 2',
-		label: 'Line 2',
-		data: [
-			{ date: new Date( '2024-01-01' ), value: 1, label: 'Jan 1' },
-			{ date: new Date( '2024-01-02' ), value: 2, label: 'Jan 2' },
-			{ date: new Date( '2024-01-03' ), value: 1.5, label: 'Jan 3' },
-			{ date: new Date( '2024-01-04' ), value: 2.5, label: 'Jan 4' },
-			{ date: new Date( '2024-01-05' ), value: 3, label: 'Jan 5' },
-		],
-	},
-];
-
-const pieData = [
-	{
-		label: 'Windows',
-		value: 80000,
-		valueDisplay: '80K',
-		percentage: 2,
-	},
-	{
-		label: 'MacOS',
-		value: 30000,
-		valueDisplay: '30K',
-		percentage: 5,
-	},
-	{
-		label: 'Linux',
-		value: 22000,
-		valueDisplay: '22K',
-		percentage: 1,
-	},
-];
-
-const GridComponent = ( { children } ) => {
-	return (
-		<div style={ { display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', gap: '1rem' } }>
-			{ children }
-		</div>
-	);
-};
-
-export const Default: Story = {
-	render: () => (
-		<ThemeProvider>
-			<GridComponent>
-				<LineChart
-					data={ lineSampleData }
-					width={ 400 }
-					height={ 300 }
-					withGradientFill={ false }
-					withLegendGlyph={ false }
-				/>
-				<BarChart data={ sampleData } width={ 400 } height={ 300 } />
-				<PieSemiCircleChart data={ pieData } width={ 400 } label="Pie Chart" note="Default Theme" />
-			</GridComponent>
-		</ThemeProvider>
-	),
-};
-
-export const JetpackTheme: Story = {
-	render: () => (
-		<ThemeProvider theme={ jetpackTheme }>
-			<GridComponent>
-				<LineChart
-					data={ lineSampleData }
-					width={ 400 }
-					height={ 300 }
-					withGradientFill={ false }
-					withLegendGlyph={ false }
-				/>
-				<BarChart data={ sampleData } width={ 400 } height={ 300 } />
-				<PieSemiCircleChart data={ pieData } width={ 400 } label="Pie Chart" note="Jetpack Theme" />
-			</GridComponent>
-		</ThemeProvider>
-	),
-};
-
-export const WooTheme: Story = {
-	render: () => (
-		<ThemeProvider theme={ wooTheme }>
-			<GridComponent>
-				<LineChart
-					data={ lineSampleData }
-					width={ 400 }
-					height={ 300 }
-					withGradientFill={ false }
-					withLegendGlyph={ false }
-				/>
-				<BarChart data={ sampleData } width={ 400 } height={ 300 } />
-				<PieSemiCircleChart data={ pieData } width={ 400 } label="Pie Chart" note="Woo Theme" />
-			</GridComponent>
-		</ThemeProvider>
-	),
-};
-
-export const CustomTheme: Story = {
-	render: () => (
-		<ThemeProvider
-			theme={ {
-				colors: [ '#073B3A', '#0B6E4F', '#08A045', '#6BBF59', '#DDB771' ],
-				seriesLineStyles: [
-					{
-						strokeWidth: 1,
-						strokeDasharray: '8 8',
-						strokeLinecap: 'square',
-					},
-					{
-						strokeDasharray: '5 8',
-						strokeWidth: 2,
-						strokeLinecap: 'square',
-					},
-				],
-				gridStyles: {
-					stroke: '#ffe3e3',
-					strokeWidth: 2,
-				},
-			} }
-		>
-			<GridComponent>
-				<LineChart
-					data={ lineSampleData }
-					width={ 400 }
-					height={ 300 }
-					withGradientFill={ false }
-					withLegendGlyph={ false }
-				/>
-				<BarChart data={ sampleData } width={ 400 } height={ 300 } />
-				<PieSemiCircleChart
-					data={ pieData }
-					width={ 400 }
-					label="Pie Chart"
-					note="Custom Color Theme"
-				/>
-			</GridComponent>
-		</ThemeProvider>
-	),
-};
